backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.23
python-multipart>=0.0.9
google-generativeai>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.6.0
